Basic Algorithm and SVG Output
This change implements the basic algorithm for embedding a unit disk
contact graph as described in the paper.

The program reads its input from a file given as command argument and
writes the outcome to two different files with hardcoded names.

One of them is a text file with info dumped from the internal
representation, the other is a graphical SVG.