Failure Indication
One of the main features of the program is (or should be) that it not
just attempts to create an appropriate graph representation, but also
that it reports when it cannot do so due to space constraints.

With this change, vertices that cannot be placed in the plane under UDCR
constraints are marked (in text for text output, crimson fill-color for
SVG output).