Classification of arbitrary input graphs
There is now a mechanism in place that can recognize whether an
arbitrary input graph is an acceptable graph type.

It can also pick out the spine from among the graph definition.

The implementation is yet untested and is missing the conversion to a
usable DiskGraph representation that can be used by the algorithms.