Dynamic Problem Embedder infrastructure
The new `class DynamicProblemEmbedder` ties together the dynamic
programming implementation to make it accessible to the main program.

The new option `dynamic-program` for the configuration key `--algorithm`
allows the user to embed lobsters with the newest implementation.

Not all solvable cases can be solved due
to problems yet to be identified.