Enumerate: infrastructure
Adds a `class Lobster` similar to `Caterpillar` for a very slim
representation of essential lobster characteristics.

Adds a `class Enumerate` and supporting classes which will generate
lobsters of increasing size, evaluate them with heuristic and
dynamic programming embedding algorithms and collect statistics.

The tests and implementation are barebones, to be done.