Dynamic Program: Domination Heuristic
Partial dynamic programming problems no longer enter consideration if
we already have considered an instance which is strictly easier to
solve.

We determine this by considering the blocked spaces of their fundaments.
If, everything else being equal, the fundament of one problem does not
block any spaces blocked in the other fundament, it "dominates" the
other fundament and it is sufficient to consider this problem only.