Batch Size
The SVG output now supports output in batches, where one output file
will be closed after a certain number of output graphs have been written
to it. The next graphs will then be written to the next file in the
batch.

This change comes with configuration options and some restructuring of
the SVG module.