Stats: Append
Even if the program runs in single mode (not benchmark), it will now
output one CSV line to the stats file (if configured).

While the benchmark overwrites the whole file every time, for a single
trial, the line is appended to the existing file.

If the file does not exist, automatically write the header line.