Improved Usability of Coord and Dir
The two basic types Coord and Dir can now be added with operator+.

Neighbor coordinates of distance 1 to 2 now have dedicated functions in
the geometry module, making their determination more explicit.

The function step, which formerly belonged to the Grid class, also
switched to the geometry module, as did the vec function for Euclidean
coordinates.