Logging: Configuration, Silent Mode
The appropriate logging mechanism is now constructed at the start of
the program.

To disable logging, it now offers the `silent` log level, replacing the
log mode `none`.

Improved the log implementation such that it no longer does string
formatting if the log message is suppressed due to the level.