Dynamic Programming: Store Fundament instead of Solution
Rectified the stopgap measure where every dynamic programming problem
used to store the whole partial solution up to that point, which caused
a lot of overhead.

Now, a dynamic problem only keeps info about the disk that was embedded
in that step and a reference to the parent for previous steps.