Dynamic Program: Decision Mode
If no output file is specified while running the dynamic program,
the program will now refrain from keeping all the parent problems
in memory which would allow us to reconstruct the embedding solution.

Instead, we only keep the result of the decision problem, whether the
embedding is possible or not.

This may yield performance improvements.