Dynamic Program: Dominance Runtime Fix
The previous implementation used one "closed" set for all previously
encountered signatures, from which the range of signatures to check
against would be determined in time log(n), logarithmic in the total
depth of the problem.

The fixed version uses one closed set for each problem depth level,
at which the number of signatures is constant. The level selection is
also constant. This brings the implementation in line with the
expectation of linear run time.